Add transition data for Tokyo
Tokyo had Daylight Saving Time from 1948 to 1951 and some transitions
occurred on "the 25th hour of the second Saturday of September", e.g. on
1 AM on the Sunday following the second Saturday in September.

This unusual phrasing is reflected in the raw source data, but not in
the compiled TZif binaries; these tests are useful regression tests in
the event that switch to a source of data other than `zic`.