Add coverage for C extensions
The hack in the setup.py file is there because `gcov` generates a
*required* .gcno file at build time at the location of the object files.
In our case, these object files only ever exist in an ephemeral
temporary build directory that pip has moved copied all of our files
into, for build isolation purposes.

This allows you to copy those into some location, specified by the
GCNO_TARGET_DIR directory, at build time. Maybe in the future we either
use a different coverage generation tool or figure out a way to relocate
this .gcno file automatically, without hacking it into setup.py.

Additionally, I should note that the ideal target for this would
actually be {envtmpdir}, because that's cleared before the environment
is run each time, but unfortunately it's cleared *after* the install
command, not before, so anything we copy in there as part of the build
will get deleted. There is currently no way to insert a command before
the build occurs to establish our own version of envtmpdir (...with
blackjack ...and hookers...).