Use zero-transition zones in TZStr tests
These tests initially tried to use zero-transition zones, but that edge
case was not well-supported in the codebase. Now that those problems are
fixed, we can switch over to using zero-transition zones by default.

This commit also switches over to putting most of the constant parts of
the TZif file into a dedicated class attribute, populated at class setup
time.