Add support for rule-based transitions
In Version 2+ files, after the last transition in a file, a string based
on the POSIX TZ environment variable is used to determine transitions
going forward indefinitely.

This brings the C extension to feature parity with the Python module,
and now all tests enabled for the Python version are enabled for the C
extension as well.