Add support for zones with prefix abbreviations
Evidently to save space, if one abbreviation is the same as another
abbreviation but with a prefix (e.g. AHST and HST), the ttinfo with the
shorter abbreviation will point to wherever the subset string starts.
This makes it so that the strategy of eagerly breaking up the
abbreviation characters into strings won't work nearly as well, so
instead we lazily build the abbreviation strings.