Fix version 1 conversion routine
There were several errors here that were not caught because file_size
was always much larger than the actual size of the file, so these files
were being included whole and unmodified. The results were likely not
affected much because by changing the header to a Version 1 header, the
Version 2 file was ignored.