Support TZif\0 instead of TZif1
According to RFC 8536 §3.1, the version indicator for version 1 files is
'\0' (0x00) (NUL), rather than '1' (0x31) — which makes sense, since the
version number was likely added using one of the reserved header bytes.

This version of the code will support `0x00` and also any erroneous
files with `TZif1` as the header (though it appears that there was never
a version of zic that would generate files with that version number).