Make the pure Python implementation a fallback
This seems like it would be simple, but in order to test this we need to
import the module multiple times, which reveals all kinds of problems
with test.support.import_fresh_module, and some problems with the
implementation of our tests.

Here we move the various workarounds and test support into a dedicated
test support submodule, so that they can be used in both the unit and
proper tests.

Additionally, it seems like it's not possible to pickle something where
the name refers to a different module when imported fresh (e.g. from
sys.modules). This workaround makes sure that the class we are testing
is always the one you get from `import zoneinfo`.