Fix fold detection behavior for early transitions
When the gap between the timestamp being tested and the most recent
transition is too big to fit in a C integer, the Python implementation
will fail.

This was discovered from testing with hypothesis on Ubuntu (which has
has some super early transition in at least Africa/Abidjan for some
reason).

Adding a determinitsic test for this required reworking the arbitrary
zone generator a bit to allow for timestamp values outside the range of
valid datetimes.