Fix fold handling for one-transition zones
Prior to this change, when a zone has a single transition (particularly
to a zone with no DST), the fold handling logic in the C extension was
broken, because the logic for detecting how to handle the first offset
was incorrectly based on the number of transitions rather than the
number of ttinfos.

This bug was discovered by running the property tests over and over
again hundreds of times; the minimal reproducing case has also been
added to the unit tests.