Add pytest-randomly and pytest-xdist
pytest-randomly randomizes the order of the tests on each run, which
will help us to avoid accidental dependencies on the run order of the
tests.

Installing pytest-xdist alone has no effect on the tests, but it does
allow you to do things like `tox -e py38 -- -n auto` to run the tests in
parallel.