Add support for datetime.time
Per the documentation, `.utcoffset()`, `.dst()` and `.tzname()` should
return None when the offset is unknown; for anything with transitions,
a datetime.time() is not enough information to resolve the offset.

For fixed offset zones, we know the right response - though I believe
this will not correctly handle a "permanent DST" zone like
"XST0XDT,0/0,J365/25", which looks like it has transitions but doesn't.

Luckily, that sort of thing doesn't exist in the time zone database, and
even if it were added "can't get the offset from datetime.time" is not a
big deal.