Include right/ and posix/ zones in property tests
right/ and posix/ should be mirrors of the primary database, with the
same tree of time zones below them as the main directory. Since these
are zones likely to produce edge cases, but available_timezones() now longer
includes them, we can manually add them back in for these purposes.

The posixrules file is deliberately left out, here, because it's
intended to always be a link to another zone.