Add support for PyPy
This skips building the C extension and only uses the pure Python
version when built on PyPy. This is probably the right thing to do on
PyPy anyway, since PyPy needs to use a compatibility layer for the C
extension, and (as far as I know) doesn't get the same sort of speedup.