Add support for PyPy
This skips building the C extension and only uses the pure Python
version when built on PyPy. This is probably the right thing to do on
PyPy anyway, since PyPy needs to use a compatibility layer for the C
extension, and that slows things down considerably.

Running the benchmark suite, the C extension is consistently 10x slower
than the pure Python implementation on PyPy.