Build wheel for PyPy
This is a tricky situation, because the PyPy wheel is special — it does
not contain the C extension. We do not want this being picked up by
anything except pypy.

Unfortunately, after pip version 19.3.1 (and until at least version
20.1.1), pip will not pick up the `pp3-none-any.whl` tag (see
https://github.com/pypa/pip/issues/8347 ), so pypy users with a recent
version of pip will get the sdist (which is our only alternative really
anyway). We'll ship these wheels anyway, since this seems to be an issue
with pip, and this is indeed the proper tag for PyPy 3 wheels.