Remove setup-cfg-fmt
It was nice while it lasted, but there's no option to disable the
automatic addition of Python 3.9 to the trove classifiers (it would be
inappropriate to include it or to exclude it from `python_requires`).

That plus the fact that it deletes all comments intentionally makes this
a tougher trade-off than I'd like, so we'll just have to be diligent
about the formatting in the future.