Pin hypothesis to <5.43.0
Hypothesis added a `timezones` method, which has a dependency on
`backports.zoneinfo` and that messes up the import hacks we're doing to
test both the C and the Python versions of the zoneinfo module.

For now we'll pin it until we come up with a solution that allows us to
import hypothesis on the newest versions.