Update tzdata to 2021a
Historical updates to Australia/Sydney would break the tests starting
with 2020e. This leads to a discrepancy between how the test suite
fairs with the latest `tzdata` installed and the pinned data we've
vendored.

The relevant commit on github.com/eggert/tz is 1aca4cad3692063cd5cb571883ccf6d04b3c2991

https://mm.icann.org/pipermail/tz/2020-November/029496.html

This updates the pinned data and our tests.