Pin Github Actions runner versions
The latest runners no longer support Python 3.6, and presumably this
trend will continue and support for 3.7 and 3.8 will also eventually be
dropped.

Since this is a backport and its whole raison d'être is to support older
versions of Python, we are much more concerned with testing on older
versions than on the latest runners, so we will pre-emptively pin all
the runners to versions that work with 3.6, 3.7 and 3.8.

In this commit, we also specify that "pypy" means "pypy3.8".