Pin importlib_metadata
We need an old version of hypothesis because of 4845961fba397deacaf47a38689a6298af9c63bb

But the older versions of hypothesis throw an error when used on
importlib_metadata >= 5.0.0 (see
https://github.com/HypothesisWorks/hypothesis/issues/3473), so we'll pin
importlib_metadata.

I don't think we need to worry too much about trying to keep up with
newer versions of these libraries since we're unlikely to use this
namespace to backport new zoneinfo features.